requirements.txt
bidict==0.23.1
blinker==1.9.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
dnspython==2.7.0
eventlet==0.40.2
filelock==3.18.0
Flask==3.1.1
flask-cors==6.0.1
Flask-SocketIO==5.5.1
fonttools==4.59.0
fsspec==2025.7.0
greenlet==3.2.3
h11==0.16.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
numpy==2.2.6
opencv-python==4.12.0.88
packaging==25.0
pandas==2.3.1
pillow==11.3.0
psutil==7.0.0
py-cpuinfo==9.0.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-engineio==4.12.2
python-socketio==5.13.0
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
scipy==1.16.1
setuptools==80.9.0
simple-websocket==1.1.0
six==1.17.0
sympy==1.14.0
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.171
ultralytics-thop==2.0.14
urllib3==2.5.0
Werkzeug==3.1.3
wsproto==1.2.0
